+++ src/main/java/org/nextme/point_service/point/domain/Point.java
@@ -1,4 +1,4 @@
-package org.nextme.point_service.domain;
+package org.nextme.point_service.point.domain;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
